requirements.txt
langchain
langchain-openai
google-cloud-bigquery
langchain-google-genai
beautifulsoup4
Flask